# Refactor columns introspection query to make it faster