--- a/test/support/sql_counter_sqlserver.rb
+++ b/test/support/sql_counter_sqlserver.rb
@@ -14,6 +14,7 @@ module ARTest
 
     ignored_sql = [
       /INFORMATION_SCHEMA\.(TABLES|VIEWS|COLUMNS|KEY_COLUMN_USAGE)/im,
+      /sys.columns/i,
       /SELECT @@version/,
       /SELECT @@TRANCOUNT/,
       /(BEGIN|COMMIT|ROLLBACK|SAVE) TRANSACTION/,