@BenMcH 2021-01-10
@wpolicarpo See my issue created here: #850 

This PR seems to have created a regression for us